Handle cite spaces properly
The \cite command is often used with a preceeding space, succeeded by punctuation. However, since the tool
translates this to just a space, Grammarly will complain about there being a dangling space before the punctuation.

This commit's changes can be visualised with an example:
Before: "The dog chased the cat \cite{overleaf}." -> "The dog chased the cat ."
After: "The dog chased the cat \cite{overleaf}." -> "The dog chased the cat."