Removed extra \cite
As mentioned in robindijkhof#70 there is a bug with the ´\cite´ regular expression. With this PR I removed the one that has is more complex and also selects the ",." and then subtitutes it.
Here is an visualization of the wrong expression..

Hope you can merge this into master in time to tomorrows release.